# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Thesis Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /theses/{id}/files:
    get:
      operationId: listThesisFiles
      tags: [files]
      summary: List files for a thesis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ThesisFile"
    post:
      operationId: uploadThesisFile
      tags: [files]
      summary: Upload a file to a thesis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThesisFileInput"
      responses:
        "201":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThesisFile"
  /files/{id}:
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /theses/{id}/reviews:
    get:
      operationId: listThesisReviews
      tags: [reviews]
      summary: List reviews for a thesis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}:
    get:
      operationId: getReview
      tags: [reviews]
      summary: Get review by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
    patch:
      operationId: updateReview
      tags: [reviews]
      summary: Update a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewUpdate"
      responses:
        "200":
          description: Updated review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /theses/{id}/grades:
    get:
      operationId: listThesisGrades
      tags: [grades]
      summary: List grades for a thesis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Grades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Grade"
    post:
      operationId: createGrade
      tags: [grades]
      summary: Submit a grade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeInput"
      responses:
        "201":
          description: Grade submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grade"
  /grades/{id}:
    patch:
      operationId: updateGrade
      tags: [grades]
      summary: Update a grade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeUpdate"
      responses:
        "200":
          description: Updated grade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grade"
  /defenses:
    get:
      operationId: listDefenses
      tags: [defenses]
      summary: List defense sessions
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Defense sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DefenseSession"
    post:
      operationId: createDefense
      tags: [defenses]
      summary: Schedule a defense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DefenseInput"
      responses:
        "201":
          description: Scheduled defense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefenseSession"
  /defenses/{id}:
    get:
      operationId: getDefense
      tags: [defenses]
      summary: Get defense session by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Defense session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefenseSession"
    patch:
      operationId: updateDefense
      tags: [defenses]
      summary: Update defense session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DefenseUpdate"
      responses:
        "200":
          description: Updated defense session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefenseSession"
    delete:
      operationId: deleteDefense
      tags: [defenses]
      summary: Delete defense session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [student, supervisor, reviewer, committee_member, admin]
        faculty:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        phoneNumber:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - createdAt
    RegisterInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [student, supervisor, reviewer, committee_member, admin]
        faculty:
          type: string
        department:
          type: string
        phoneNumber:
          type: string
      required:
        - email
        - password
        - firstName
        - lastName
        - role
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
...
    ThesisFile:
      type: object
      properties:
        id:
          type: integer
        thesisId:
          type: integer
        fileName:
          type: string
        fileUrl:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
        uploadedBy:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - thesisId
        - fileName
        - fileUrl
        - fileType
        - fileSize
        - uploadedBy
        - createdAt
    ThesisFileInput:
      type: object
      properties:
        fileName:
          type: string
        fileUrl:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
      required:
        - fileName
        - fileUrl
        - fileType
        - fileSize
    Review:
      type: object
      properties:
        id:
          type: integer
        thesisId:
          type: integer
        reviewerId:
          type: integer
        content:
          type: string
        recommendation:
          type: string
          enum: [approve, reject, revise]
        isPublished:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
        reviewer:
          $ref: "#/components/schemas/User"
      required:
        - id
        - thesisId
        - reviewerId
        - content
        - recommendation
        - isPublished
        - createdAt
        - updatedAt
        - reviewer
    ReviewInput:
      type: object
      properties:
        content:
          type: string
        recommendation:
          type: string
          enum: [approve, reject, revise]
        isPublished:
          type: boolean
      required:
        - content
        - recommendation
    ReviewUpdate:
      type: object
      properties:
        content:
          type: string
        recommendation:
          type: string
          enum: [approve, reject, revise]
        isPublished:
          type: boolean
    Grade:
      type: object
      properties:
        id:
          type: integer
        thesisId:
          type: integer
        graderId:
          type: integer
        value:
          type: number
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
        grader:
          $ref: "#/components/schemas/User"
      required:
        - id
        - thesisId
        - graderId
        - value
        - createdAt
        - grader
    GradeInput:
      type: object
      properties:
        value:
          type: number
        comment:
          type: string
      required:
        - value
    GradeUpdate:
      type: object
      properties:
        value:
          type: number
        comment:
          type: string
    DefenseSession:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        scheduledAt:
          type: string
        location:
          type: ["string", "null"]
        roomOrLink:
          type: ["string", "null"]
        thesisIds:
          type: array
          items:
            type: integer
        committeeIds:
          type: array
          items:
            type: integer
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - scheduledAt
        - thesisIds
        - committeeIds
        - createdAt
    DefenseInput:
      type: object
      properties:
        title:
          type: string
        scheduledAt:
          type: string
        location:
          type: string
        roomOrLink:
          type: string
        thesisIds:
          type: array
          items:
            type: integer
        committeeIds:
          type: array
          items:
            type: integer
        notes:
          type: string
      required:
        - title
        - scheduledAt
    DefenseUpdate:
      type: object
      properties:
        title:
          type: string
        scheduledAt:
          type: string
        location:
          type: string
        roomOrLink:
          type: string
        thesisIds:
          type: array
          items:
            type: integer
        committeeIds:
          type: array
          items:
            type: integer
        notes:
          type: string
...