# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Thesis Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  role: UserRole;
  /** @nullable */
  faculty?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}